Move serial resim tests to single test
