+++ simulator/src/resim/mod.rs
@@ -555,3 +555,74 @@ pub fn db_upsert_epoch(epoch: Epoch) -> Result<(), Error> {
 
     Ok(())
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn test_no_value() {
+        let mut out = std::io::stdout();
+        assert!(Reset {}.run(&mut out).is_ok());
+        let new_account = NewAccount {
+            network: None,
+            manifest: None,
+            trace: false,
+        };
+        assert!(new_account.run(&mut out).is_ok());
+        let cmd = Show { address: None };
+        assert!(cmd.run(&mut out).is_ok());
+    }
+
+    fn test_pre_process_manifest() {
+        temp_env::with_vars(
+            vec![
+                (
+                    "faucet",
+                    Some("system_sim1qsqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqpql4sktx"),
+                ),
+                (
+                    "xrd",
+                    Some("resource_sim1qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqzqu57yag"),
+                ),
+            ],
+            || {
+                let manifest = r#"CALL_METHOD ComponentAddress("${  faucet  }") "free";\nTAKE_ALL_FROM_WORKTOP ResourceAddress("${xrd}") Bucket("bucket1");\n"#;
+                let after = r#"CALL_METHOD ComponentAddress("system_sim1qsqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqpql4sktx") "free";\nTAKE_ALL_FROM_WORKTOP ResourceAddress("resource_sim1qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqzqu57yag") Bucket("bucket1");\n"#;
+                assert_eq!(Run::pre_process_manifest(manifest), after);
+            },
+        );
+    }
+
+    fn test_set_default_account_validation() {
+        let mut out = std::io::stdout();
+        let private_key = Secp256k1PrivateKey::from_hex(
+            "6847c11e2d602548dbf38789e0a1f4543c1e7719e4f591d4aa6e5684f5c13d9c",
+        )
+        .unwrap();
+        let public_key = private_key.public_key().to_string();
+
+        let make_cmd = |key_string: String| {
+            return SetDefaultAccount {
+                component_address: SimulatorComponentAddress::from_str(
+                    "account_sim1c9yeaya6pehau0fn7vgavuggeev64gahsh05dauae2uu25njk224xz",
+                )
+                .unwrap(),
+                private_key: key_string,
+                owner_badge: SimulatorNonFungibleGlobalId::from_str(
+                    "resource_sim1ngvrads4uj3rgq2v9s78fzhvry05dw95wzf3p9r8skhqusf44dlvmr:#1#",
+                )
+                .unwrap(),
+            };
+        };
+
+        assert!(make_cmd(private_key.to_hex()).run(&mut out).is_ok());
+        assert!(make_cmd(public_key.to_string()).run(&mut out).is_err());
+    }
+
+    #[test]
+    fn serial_resim_command_tests() {
+        test_no_value();
+        test_pre_process_manifest();
+        test_set_default_account_validation();
+    }
+}
